28/requirements_develop.txt
-r requirements.txt
django-debug-toolbar==2.2
Faker==4.1.1
model-bakery==1.1.1
pytest==5.4.3
pytest-django==3.9.0
